+++ .mocharc.js
@@ -5,4 +5,5 @@ module.exports = {
   exit: true,
   slow: 100,
   timeout: 90000,
+  require: ['ts-node/register/transpile-only'],
 };