+++ src/ErrorWithCode.ts
@@ -1,0 +1,9 @@
+export class ErrorWithCode extends Error {
+  public code: string;
+
+  public constructor(message: string, code: string) {
+    super(message);
+
+    this.code = code;
+  }
+}